Split LSP into multiple files
