+++ linefeed-lsp/src/capabilities.rs
@@ -1,0 +1,46 @@
+use tower_lsp::lsp_types::*;
+
+/// Semantic token types legend used by the LSP server
+pub const LEGEND_TYPE: &[SemanticTokenType] = &[
+    SemanticTokenType::FUNCTION,
+    SemanticTokenType::VARIABLE,
+    SemanticTokenType::STRING,
+    SemanticTokenType::COMMENT,
+    SemanticTokenType::NUMBER,
+    SemanticTokenType::KEYWORD,
+    SemanticTokenType::OPERATOR,
+    SemanticTokenType::PARAMETER,
+];
+
+/// Build the server capabilities for initialization
+pub fn build_server_capabilities() -> ServerCapabilities {
+    ServerCapabilities {
+        text_document_sync: Some(TextDocumentSyncCapability::Kind(
+            TextDocumentSyncKind::FULL,
+        )),
+        semantic_tokens_provider: Some(
+            SemanticTokensServerCapabilities::SemanticTokensRegistrationOptions(
+                SemanticTokensRegistrationOptions {
+                    text_document_registration_options: TextDocumentRegistrationOptions {
+                        document_selector: Some(vec![DocumentFilter {
+                            language: Some("lf".to_string()),
+                            scheme: Some("file".to_string()),
+                            pattern: None,
+                        }]),
+                    },
+                    semantic_tokens_options: SemanticTokensOptions {
+                        work_done_progress_options: WorkDoneProgressOptions::default(),
+                        legend: SemanticTokensLegend {
+                            token_types: LEGEND_TYPE.into(),
+                            token_modifiers: vec![],
+                        },
+                        range: Some(true),
+                        full: Some(SemanticTokensFullOptions::Bool(true)),
+                    },
+                    static_registration_options: StaticRegistrationOptions::default(),
+                },
+            ),
+        ),
+        ..ServerCapabilities::default()
+    }
+}
